cleanup

+++ Ookii.Dialogs.Wpf/VistaFileDialog.cs
@@ -683,7 +683,7 @@ namespace Ookii.Dialogs.Wpf
                 int result = dialog.Show(hwndOwner);
                 if( result < 0 )
                 {
-                    if( (uint)result == (uint)HRESULT.ERROR_CANCELLED )
+                    if( result == (int)HRESULT.ERROR_CANCELLED )
                         return false;
                     else
                         throw System.Runtime.InteropServices.Marshal.GetExceptionForHR(result);
